# services/tracker/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorTaskAttempt": {
        "properties": {
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_fingerprint": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Error Fingerprint",
            "type": "string"
          },
          "error_message": {
            "maxLength": 4000,
            "title": "Error Message",
            "type": "string"
          },
          "error_message_truncated": {
            "title": "Error Message Truncated",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "const": "error",
            "default": "error",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "attempt_id",
          "created_at",
          "error_message",
          "error_message_truncated",
          "error_fingerprint"
        ],
        "title": "ErrorTaskAttempt",
        "type": "object"
      },
      "EvaluationTaskAttempt": {
        "properties": {
          "agent_caused_exit_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentCausedExitReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instance_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance Id"
          },
          "kind": {
            "const": "evaluation",
            "default": "evaluation",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "attempt_id",
          "created_at",
          "instance_id",
          "agent_caused_exit_reason"
        ],
        "title": "EvaluationTaskAttempt",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SingleTaskResponse": {
        "properties": {
          "agent_caused_exit_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Caused Exit Reason"
          },
          "error_message": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "evaluation_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Result"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "status",
          "started_at",
          "finished_at",
          "error_message",
          "evaluation_result",
          "agent_caused_exit_reason"
        ],
        "title": "SingleTaskResponse",
        "type": "object"
      },
...
      "TaskArtifactsResponse": {
        "properties": {
          "agent_output_expires_in": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Output Expires In"
          },
          "agent_output_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Output Url"
          },
          "cloudwatch_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudwatch Url"
          }
        },
        "required": [
          "cloudwatch_url",
          "agent_output_url",
          "agent_output_expires_in"
        ],
        "title": "TaskArtifactsResponse",
        "type": "object"
      },
      "TaskAttemptsResponse": {
        "properties": {
          "attempts": {
            "items": {
              "discriminator": {
                "mapping": {
                  "error": "#/components/schemas/ErrorTaskAttempt",
                  "evaluation": "#/components/schemas/EvaluationTaskAttempt",
                  "execution": "#/components/schemas/ExecutionTaskAttempt"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorTaskAttempt"
                },
                {
                  "$ref": "#/components/schemas/EvaluationTaskAttempt"
                },
                {
                  "$ref": "#/components/schemas/ExecutionTaskAttempt"
                }
              ]
            },
            "title": "Attempts",
            "type": "array"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "attempts",
          "total_count"
        ],
        "title": "TaskAttemptsResponse",
        "type": "object"
      },
...
      "TaskLogAttemptsResponse": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/TaskLogAttempt"
            },
            "title": "Attempts",
            "type": "array"
          },
          "current_attempt_id": {
            "title": "Current Attempt Id",
            "type": "string"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "attempts",
          "current_attempt_id",
          "next_cursor"
        ],
        "title": "TaskLogAttemptsResponse",
        "type": "object"
      },
...
      "TaskLogEventsResponse": {
        "additionalProperties": false,
        "properties": {
          "attempt_id": {
            "title": "Attempt Id",
            "type": "string"
          },
          "direction": {
            "enum": [
              "forward",
              "backward"
            ],
            "title": "Direction",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TaskLogEvent"
            },
            "title": "Events",
            "type": "array"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "newer_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newer Cursor"
          },
          "older_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Older Cursor"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          }
        },
        "required": [
          "attempt_id",
          "is_current",
          "is_active",
          "task_status",
          "direction",
          "events",
          "older_cursor",
          "newer_cursor"
        ],
        "title": "TaskLogEventsResponse",
        "type": "object"
      },
      "TaskStatus": {
        "enum": [
          "PENDING",
          "BUILDING",
          "IN_PROGRESS",
          "EVALUATING",
          "STOPPED",
          "FINISHED",
          "ERROR"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DescopeAccessKey": {
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      },
      "DescopeBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/benchmarks/{benchmark_id}/tasks/{task_id}": {
      "get": {
        "description": "Fetch a single task's status + evaluation result for the SingleTask page.",
        "operationId": "get_single_task",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Single Task"
      }
    },
    "/benchmarks/{benchmark_id}/tasks/{task_id}/artifacts": {
      "get": {
        "description": "CloudWatch URL + presigned URL for the agent's output tarball, for the SingleTask page.",
        "operationId": "get_task_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskArtifactsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Task Artifacts"
      }
    },
...
    "/benchmarks/{benchmark_id}/tasks/{task_id}/attempts": {
      "get": {
        "description": "Return every persisted task outcome, newest first.",
        "operationId": "get_task_attempts",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskAttemptsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Task Attempts"
      }
    },
...
    "/benchmarks/{benchmark_id}/tasks/{task_id}/log-attempts": {
      "get": {
        "description": "Page exact CloudWatch streams for every retry of a task.",
        "operationId": "get_task_log_attempts",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLogAttemptsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Task Log Attempts"
      }
    },
    "/benchmarks/{benchmark_id}/tasks/{task_id}/log-attempts/{attempt_id}/events": {
      "get": {
        "description": "Page one attempt's events; forward cursors efficiently tail active work.",
        "operationId": "get_task_log_attempt_events",
        "parameters": [
          {
            "in": "path",
            "name": "benchmark_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Benchmark Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "attempt_id",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[0-9a-f]+$",
              "title": "Attempt Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "forward",
              "enum": [
                "forward",
                "backward"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskLogEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "DescopeBearer": []
          },
          {
            "DescopeAccessKey": []
          }
        ],
        "summary": "Get Task Log Attempt Events"
      }
    },
...
  }
}